Speed up tests yet another 45 seconds
This changes the security groups extension tests to use the Controllers
directly, bypassing the slow wsgi setup process. This is part 2 of this
patch and contains the majority of the changes. The code required more
changes than the other test suite speedups because of how the previous
code was written and as a result, the diff is kind of long. As a bonus,
there is a net removal of almost 150 lines.

Change-Id: I7893cf16092ce58860607fd4ab2df10550fe3270