Remove install_requires processing.
Loading install_requires with the contents of pip-requires
isn't getting us any real beneift and is causing issues.

a) It can conflict with installing nova into an environment
   where deps have been installed from packages (devstack)
b) It breaks the ability to use -e git urls in pip-requires
   which we want to start using for python-novaclient and
   python-keystoneclient
c) It causes spurious network traffic when we're trying to
   test things.

At the same time, since we are not expecting anyone to
install nova from setup.py for production, the normal benefit
of the feature is not needed.

Change-Id: I402f975c652dc1c2b54b0c362b8abbb6886f8844