Add missing scripts to setup.py (lp#917676)
If distutils-extra is installed, all scripts are automatically included
in the tarball. We should explicitly include them in setup.py to avoid
confusion.

As an example, the 2011.3 tarballs appear to have been generated with
distutils-extra but the 2011.3.1/2012.1 tarballs are now being
generated in a venv. So, these newer tarballs are missing scripts which
were included in 2011.3.

Change-Id: I2a85418b84b93e041f6c7d31833b09acdcb29daa