Fix issues with security group auths without ports.
Fix bug 946427.

There was a bug where a security group would get completely opened in
cases where only icmp, udp, or tcp should be opened.  For example, any
of the following three commands would result in opening everything:

    euca-authorize -P icmp -o test-ports test-ports
    euca-authorize -P tcp -o test-ports test-ports
    euca-authorize -P udp -o test-ports test-ports

This patch resolves this and these commands now only open the protocol
that was specified.  Unit tests have been added to verify the fix and
also verify that this only works when a source group is specified.
While the bug was originally reported against the EC2 API, the same
updates and similar unit tests have gone in to the equivalent code for
the OpenStack API.

Change-Id: I4c87c5f5f4ccee60c6c16da4e659d73ab3f4a34f