Sort results from describe_instances in EC2 API.
Fix bug 827619.

This bug pointed out that EC2 sorts the results of of describe_instances
by the launch time.  Make our implementation of the EC2 API behave the same
way.  Previously, instances coming out of the db API were sorted by the
key we wanted, but in the opposite order.  You can now specify both a
sort key and order.  The behavior is the same by default, but the EC2
API sets the parameter to do an ascending sort.

Change-Id: Ifd0bc79ad4c4c8c45809dbb1ac2dadf8abcfd4c3