This patch stops metadata from being deleted when an instance is deleted.
This patch makes it so that instance metadata, system metadata, and block
device mappings are no longer set to deleted when an instance is deleted.
It doesn't make sense to delete them just because the instance they describe
has been deleted.

This patch DOES NOT MIGRATE currently deleted metadata or block
device mappings for currently deleted instances, that means that usage
notifications for deleted instances that were deleted before this patch will
show image_meta as empty.

Fixes bug 1021830

Change-Id: I33ee402b859834e09535c7925957bee0a1d91207