Fix SQL deadlock in quota reservations
Fixes bug 1026709

The code in quota_reserve acquires SQL locks in a different order than
the code in reservation_commit/reservation_rollback. This can result in
an SQL deadlock under heavy load. Due to an (unrelated) bug in
SQLAlchemy, this can result in this exception:

ResourceClosedError: This result object does not return rows. It has
been closed automatically.

This patch reorganizes the code to always fetch (and thusly lock) the
quota_usages table before the reservations table.

Change-Id: Ia364496a996870d754094915ea0501ff19052037