Add ops to aggregate_instance_extra_specs filter.
The following operators on instance_type_extra_specs were added in
aggregate_instance_extra_specs filter. Fixes bug 1037090.

= (equal to or greater than as a number; same as vcpus case)
== (equal to as a number)
!= (not equal to as a number)
>= (greater than or equal to as a number)
<= (less than or equal to as a number)
s== (equal to as a string)
s!= (not equal to as a string)
s>= (greater than or equal to as a string)
s> (greater than as a string)
s<= (less than or equal to as a string)
s< (less than as a string)
<in> (substring)
<or> (find one of these)

If the first word in the instance_type_extra_spec is not one of the
keywords above, a default operator "s==" is used.

E.g.: ">= 5", "s== 2.1.0", "<in> gcc", and "<or> fpu <or> gpu"

Change-Id: Iee9bf68011738aab146e6311a17cb3c673d7a392