Stops compute api import at import time
There is an automatic import of compute.api at import time
whenever anything under compute is imported. This leads to all
sorts of weird dependency problems such as the one here:
https://review.openstack.org/#/c/14170/

This modifies the import to happen when compute.API() is called,
which is how we handle it for network and volume. A few imports
were also modified to make it consistent with this new style.

Change-Id: Ic748376510aa4cc48dfefc6ec36621d0ae498bfc