Make sure instance data is always refreshed
Fixes bug 1078793

When updating an instance's instance_type_id, the 'instance_type' joined
to the instance is not updated.  This updates that.  A DB API test is added
that failed before and passes now.

Also: Some cases of passing stale instance data within resize_instance()
and finish_resize() were found and addressed.

Change-Id: If335cc286a71597d3100425080e51c75aeec7a50