Fix positional arg swallow decorator
The update() method in the hosts extension is decorated with @check_host
but the function returned by decorator takes a kwarg which shadows the
'body' positional arg of the update() method.

i.e. if you call:

  controller.update(req, id, body)

then the body arg gets passed to the wrapper as its service kwarg and
never passed through to update(). We can see how the tests are all
passing body as a kwarg to avoid this.

Remove the service arg from check_host() but also from _list_hosts()
since the only place it is used is in the tests.

Change-Id: I12aab582c15c25fd35ecda005341291450c7efdc