Fix uses of service_get_all_compute_by_host
There should never be more than 1 service entry for the same topic and
host, however this method returns a list of services.  All callers of
this method except for 1 already account for there being only 1 service
entry for a particular 'compute' host and only use the first entry in the
return list.

This patch renames the DB API call to be 'service_get_by_compute_host'
and returns a single entry.  All uses of the old method are adjusted
accordingly.

Change-Id: I0e0ef62f5d2e71efe756940d9fdd98aa02fef216