Optimize network calls by moving them to api
Now that quantumv2 replaces api, there is no need for many of the
network calls to go over rpc. This patch optimizes them by moving
the implementation into the api class. Note that the old methods
are left in place to keep rpcapi compatibility.

This patch also adds api_deprecated in case anyone is using a
custom out-of-tree manager class. It is possible to set:

network_api_class=nova.network.api_deprecated.API

to get the old functionality.

Part of blueprint optimize-nova-network

Change-Id: I130908df060246e8a5f3711cf16d1c49ee3e2664