Retry floating_ip_fixed_ip_associate on deadlock.
Update the floating_ip_fixed_ip_associate method of the sqlalchemy db
API to retry if it fails because of a deadlock.  The decorator that
handles this was introduced in e53e22271ed7e7a9b919d817a8eb50a1ecce16f8.
The related bug report shows that this method could benefit from the use
of this decorator.

Fix bug 1090016.

Change-Id: I0159470e6eb2f8017bcb3e659e7b119194dda920