Translate NoMoreFloatingIps exception
When the pool of floating ips is exhausted the NoMoreFloatingIps exception
is raised but not handled properly in the floating ips extension resulting
in the exception propagating further up and causing a stack trace to be logged.

This change translates the exception explicitly into a HTTPNotFound, preserving the current
REST API behaviour. It fixes a bug where the pool name was not correctly inserted into
the message returned if allocation from a specific pool was requested.

Fixes bug 1169811

Part of blueprint no-stacktraces-in-logs

Change-Id: I8f35d25d065bb1fa709cff6f59841ac8c86658bd