xenapi: implement get_console_output for XCP/XenServer
If an administrator has enabled the logging of guest
consoles on XCP/XenServer, this enables nova to return
the last MB of those logs to the user.

The management of the logs on the server is a little
tricky, and will be sorted in a later patch.

Change was based on this previous idea:
https://review.openstack.org/#/c/17959/

DocImpact
Part of blueprint xenapi-server-log
Change-Id: I23c83bcf8c648cc2714a0c78951acc29a16d5c31