Fix info_cache and bw_usage update race
A race existed where it was possible two greenthreads to attempt to create a
new info_cache or bw_usage record for a given instance. This would result in a
DBDuplicateEntry exception.

In both of these cases, first-one-wins is an acceptable solution, so we can
swallow the exception safely.

Fixes bug 1199468

Change-Id: Icc81bb013fa9dc142f88adbe4d11cdbd350230cc