Fix deferred delete use of objects
Commit dce64683291ba2cdb5e6617e01ccc2909254acb4 added the want_objects
flag to compute_api.get to the _restore function for the V3 version
of deferred delete and the _force_delete function for the V2 version.

This changeset makes it consistent so the want_objects flag is
passed to both. The unittests are also updated accordingly. The inconsistency
was found by the integrated test_servers test which is not yet merged
for the V3 API.

Change-Id: I797f8da855fda8df823e9093cb51ef5da6bdcb0f