Fix aggregate creation/update with null or too long name
Fixes aggregate creation so an attempt to create an aggregate with
a null name results in a 400 Bad Request rather than 409 Conflict.
This was fixed for the V2 api in 43f9028f84ffec9b465bec10025ba75266c1bc22
and is fixed in this changeset for the V3 api

For both the V2 and V3 APIs adds a check to ensure that the aggregate
name is not longer than 255 characters.

Also fixes null or too long names when updating an aggregate

Fixes bug 1212240

Change-Id: I644a8a46f0b75e3717ecafec1889ce6316ca18c6