Fixes sync issue for user level resources
fixed/floating ips and networks can't be synced properly because they're
tracked per user in the quota table but they're not tied to a user so
it's not feasible to grab a count of how many are being used by any
particular user. So the quota_usages table can get into a bad state if
it sync with user level.
Sync these resources with project level, and set quota with project
level to solve the problem.

Fixes bug 1214523

Change-Id: Icd94e95fb706ee90e0737e85cfa50f05ff3c0ff8