Add mock to test-requirements
There are several changes which depend on using mock rather than mox for
unit tests:

https://review.openstack.org/#/c/42474/
https://review.openstack.org/#/c/43161/

And the general trend seems to be to move from using mock over mox for
new tests (see blueprint mox-to-mock-conversion).

And given mock is listed in the global-requirements.txt file in the
openstack/requirements project, we should require mock for nova so we
can start using it.

Change-Id: Iaa0f233c28de2f92a237ee6d089858ee10bc6d2f