Clean up object comparison routines in tests
Many object tests have the need to gracefully compare a
dict-like DB object with a NovaObject. That entails common
things like ignoring datetime timezones, etc. All of that
was duplicated in each test, but slightly tweaked. This
adds a common routine for that on the base test and makes
the other tests use it instead.

Change-Id: I6d5f0cdb8ee4d86addaec5315c55ab3a24276c9d