Ensure anti affinity scheduling works
The query for the group in the system_metadata would not correctly
filter the instances with the groups.

In addition to this change 4a03071db23196c4ac1ec0d8ea163c46887d3b08
ensured that the scheduling filter was only run once.
In cases when provisioning fails and rescheduling
takes place the filter must be invoked (if it is in the filter tool
chain).

Fixes bug: 1218878
Fixes bug: 1206502

Change-Id: I623c73635c36ffa038ae01476a25ba0c0863e536