Convert TestCases to NoDBTestCase
NoDBTestCase is faster because it skips initializing a database for
tests which don't actually talk to the database. Its also nicer
because its more explicit about the dependancies of a given test.
This is a third pass at moving test cases over to the database
lacking version, but further work is required here.

Change-Id: Ibbe86329c846e9d6df16f3bbe53a75efff6fd604
Partial-Bug: 1227473