Raise HTTPForbidden from os-floating-ips API rather than 404
The os-floating-ips extension was catching and handling a 404
and 403 in the same block and raising HTTPNotFound for both.

This change separates the 404 and 403 cases so HTTPForbidden
is returned in the 403 case.  For the 404 case, the test is
updated to raise the proper NovaException and a new test is
added for the 403 Forbidden case.

Change-Id: Id0daf2d61c6960d845d3b7980f96228fece41128