Check that volume != root device during boot by image
Currently, if an instance is booted by image and a volume is also
specified, if the volume name == the root device name, the instance
boots from the volume instead of from the image. This change adds a
check that the volume name != the root device name if an image is
specified and raises InvalidRequest if the check fails.

Closes-Bug: #1343924

Change-Id: Id967e04483b9fd12fae0bc9bee1a3e7b978b9fd1